requirements.txt
numpy==1.26.4
scipy==1.11.4
opencv-python==4.9.0.80
PyQt6==6.6.1
PyQt6-Qt6==6.6.1
PyQt6-sip==13.6.0
pandas==2.1.4
reportlab==4.0.7
python-docx==1.1.0
pillow==10.1.0
pytest==7.4.4
